Replace the rs-core/thread module with standard C++ threads.
This marks the end of support for Mingw compilation on Windows. The Mingw compiler has fallen too far behind the C++ standard; it's reached the point where it was easier to port my code to MSVC than to continue maintaining all the code that only existed to replace standard library facilities that everybody except Mingw already supports.