Fix crashes with gcc-73-cygwin
Compiler seems to be having problems resolving to_str() when we supply an ostream operator but not an explicit to_str() overload. Not sure if this is my fault or the compiler's. gcc-72-nuwen didn't have this problem back when I was using it, and this a pre-release test build from Cygwin, so it may be a compiler bug.